Change data type to double.

+++ Domain/Models/Task.cs
@@ -14,8 +14,8 @@ namespace SchedulingTool.Api.Domain.Models
         public string TaskName { get; set; } = null!;
         public int Index { get; set; }
         public int NumberOfTeam { get; set; }
-        public decimal Duration { get; set; }
-        public decimal AmplifiedDuration { get; set; }
+        public double Duration { get; set; }
+        public double AmplifiedDuration { get; set; }
         public long GroupTaskId { get; set; }
         public string? Description { get; set; }
         public string? Note { get; set; }
