+++ Resources/GroupTaskResource.cs
@@ -2,8 +2,8 @@
 
 public class GroupTaskResource
 {
-  public decimal Start { get; set; }
-  public decimal Duration { get; set; }
+  public double Start { get; set; }
+  public double Duration { get; set; }
   public string Name { set; get; } = null!;
   public string Id { set; get; } = null!;
   public string Type { set; get; } = null!;
